requirements.txt
Flask==2.2.5
tensorflow==2.16.1
opencv-python-headless==4.9.0.80
numpy==1.24.3
werkzeug==2.2.3
gunicorn==20.1.0
Pillow

